requirements.txt
streamlit==1.31.1
pandas==2.2.0
gspread
google-auth
google-auth-oauthlib
google-api-python-client
pymongo==4.6.1
python-dotenv==1.0.1
supabase>=2.0.0




